Change ACC field FuelXLap from int to float, remove comments

+++ Server/ACC_Arduino/Graphics.cs
@@ -144,7 +144,7 @@ namespace AssettoCorsaSharedMemory
         public int TCCut;
         public int EngineMap;
         public int ABS;
-        public int FuelXLap;
+        public float FuelXLap;
         public int RainLights;
         public int FlashingLights;
         public int lightsStage;
